feat: plinder root init
