+++ mypy.ini
@@ -3,5 +3,5 @@ strict = True
 ignore_missing_imports = True
 disallow_untyped_decorators = False
 
-[mypy-plinder.core._version]
+[mypy-plinder._version]
 ignore_errors = True
