+++ src/plinder/__init__.py
@@ -1,0 +1,5 @@
+# Copyright (c) 2024, Plinder Development Team
+# Distributed under the terms of the Apache License 2.0
+from ._version import _get_version
+
+__version__ = _get_version()
